@@ -1,12 +1,12 @@
 'use strict';
 
-var React = require('react'),
-    Reflux = require('reflux'),
-    Link = require('react-router').Link,
-    Navigation = require('react-router').Navigation,
-    userActions = require('../actions');
+import React from 'react';
+import Reflux from 'reflux';
+import { Link, Navigation } from 'react-router';
 
-var login = React.createClass({
+import userActions from '../actions';
+
+export default React.createClass({
   mixins: [
     Navigation,
     Reflux.listenTo(userActions.login.complete, 'onLoginSuccess'),
@@ -101,5 +101,3 @@ var login = React.createClass({
     );
   }
 });
-
-module.exports = login;
