@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   API_HOST: 'localhost',
   API_PORT: 3000,
   API_PREFIX: 'v1'
